rustc_hir_analysis

Module autoderef

Structs

Autoderef

Recursively dereference a type, considering both built-in dereferences (*) and the Deref trait. Although called Autoderef it can be configured to use the Receiver trait instead of the Deref trait.

AutoderefSnapshot 🔒

Enums

AutoderefKind

Functions

report_autoderef_recursion_limit_error