Struct rustc_hir_analysis::check::dropck::SimpleEqRelation

pub(crate) struct SimpleEqRelation<'tcx> {
    tcx: TyCtxt<'tcx>,
    param_env: ParamEnv<'tcx>,
}

This is an implementation of the TypeRelation trait with the aim of simply comparing for equality (without side-effects).

It is not intended to be used anywhere else other than here.

Fields
tcx: TyCtxt<'tcx>param_env: ParamEnv<'tcx>
Implementations

impl<'tcx> SimpleEqRelation<'tcx>

fn new(tcx: TyCtxt<'tcx>, param_env: ParamEnv<'tcx>) -> SimpleEqRelation<'tcx>