Module rustc_hir_analysis::check::region

This file builds up the ScopeTree, which describes the parent links in the region hierarchy.

For more information about how MIR-based region-checking works, see the rustc dev guide.

Structs

Context

RegionResolutionVisitor 🔒

Functions

record_var_lifetime 🔒

Records the lifetime of a local variable as cx.var_parent

region_scope_tree

Per-body region::ScopeTree. The DefId should be the owner DefId for the body; in the case of closures, this will be redirected to the enclosing function.

resolve_arm 🔒

resolve_block 🔒

resolve_expr 🔒

resolve_local 🔒

resolve_pat 🔒

resolve_stmt 🔒