rustc_hir_analysis::check::region

Function resolve_cond

fn resolve_cond<'tcx>(
    visitor: &mut ScopeResolutionVisitor<'tcx>,
    cond: &'tcx Expr<'tcx>,
)

Resolve a condition from an if expression or match guard so that it is a terminating scope if it doesn’t contain let expressions.