rustc_hir_analysis::coherence

Module inherent_impls

The code in this module gathers up all of the inherent impls in the current crate and organizes them in a map. It winds up touching the whole crate and thus must be recomputed completely for any change, but it is very cheap to compute. In practice, most code in the compiler never directly requests this map. Instead, it requests the inherent impls specific to some type (via tcx.inherent_impls(def_id)). That value, however, is computed by selecting an idea from this table.

Structs

InherentCollect 🔒

Functions

crate_incoherent_impls 🔒

crate_inherent_impls 🔒

On-demand query: yields a map containing all types mapped to their inherent impls.

crate_inherent_impls_validity_check 🔒

inherent_impls 🔒

On-demand query: yields a vector of the inherent impls for a specific type.