rustc_hir_analysis::collect

Function early_bound_lifetimes_from_generics

fn early_bound_lifetimes_from_generics<'a, 'tcx>(
    tcx: TyCtxt<'tcx>,
    generics: &'a Generics<'a>,
) -> impl Iterator<Item = &'a GenericParam<'a>>

Returns the early-bound lifetimes declared in this generics listing. For anything other than fns/methods, this is just all the lifetimes that are declared. For fns or methods, we have to screen out those that do not appear in any where-clauses etc using resolve_lifetime::early_bound_lifetimes.