rustc_hir_analysis::collect

Module predicates_of

Functions

assert_only_contains_predicates_from ğŸ”’

compute_bidirectional_outlives_predicates ğŸ”’

Opaques have duplicated lifetimes and we need to compute bidirectional outlives predicates to enforce that these lifetimes stay in sync.

const_conditions ğŸ”’

const_evaluatable_predicates_of ğŸ”’

explicit_implied_const_bounds ğŸ”’

explicit_implied_predicates_of ğŸ”’

explicit_predicates_of ğŸ”’

explicit_super_predicates_of ğŸ”’

Ensures that the super-predicates of the trait with a DefId of trait_def_id are lowered and stored. This also ensures that the transitive super-predicates are lowered.

explicit_supertraits_containing_assoc_item ğŸ”’

gather_explicit_predicates_of ğŸ”’

Returns a list of user-specified type predicates for the definition with ID def_id. N.B., this does not include any implied/inferred constraints.

implied_predicates_with_filter ğŸ”’

Ensures that the super-predicates of the trait with a DefId of trait_def_id are lowered and stored. This also ensures that the transitive super-predicates are lowered.

predicates_of ğŸ”’

Returns a list of all type predicates (explicit and implicit) for the definition with ID def_id. This includes all predicates returned by explicit_predicates_of, plus inferred constraints concerning which regions outlive other regions.

trait_explicit_predicates_and_bounds ğŸ”’

type_param_predicates ğŸ”’

Returns the predicates defined on item_def_id of the form X: Foo where X is the type parameter def_id.