Enum rustc_hir_analysis::collect::resolve_bound_vars::Scope

enum Scope<'a> {
    Binder {
        bound_vars: FxIndexMap<LocalDefId, ResolvedArg>,
        scope_type: BinderScopeType,
        hir_id: HirId,
        s: &'a Scope<'a>,
        where_bound_origin: Option<PredicateOrigin>,
    },
    Body {
        id: BodyId,
        s: &'a Scope<'a>,
    },
    ObjectLifetimeDefault {
        lifetime: Option<ResolvedArg>,
        s: &'a Scope<'a>,
    },
    Supertrait {
        bound_vars: Vec<BoundVariableKind>,
        s: &'a Scope<'a>,
    },
    TraitRefBoundary {
        s: &'a Scope<'a>,
    },
    AnonConstBoundary {
        s: &'a Scope<'a>,
    },
    Root {
        opt_parent_item: Option<LocalDefId>,
    },
}

Variants

Binder

Fields

bound_vars: FxIndexMap<LocalDefId, ResolvedArg>

We use an IndexMap here because we want these lifetimes in order for diagnostics.

scope_type: BinderScopeType

hir_id: HirId

The late bound vars for a given item are stored by HirId to be queried later. However, if we enter an elision scope, we have to later append the elided bound vars to the list and need to know what to append to.

s: &'a Scope<'a>

where_bound_origin: Option<PredicateOrigin>

If this binder comes from a where clause, specify how it was created. This is used to diagnose inaccessible lifetimes in APIT:

fn foo(x: impl for<'a> Trait<'a, Assoc = impl Copy + 'a>) {}

Declares lifetimes, and each can be early-bound or late-bound. The DebruijnIndex of late-bound lifetimes starts at 1 and it should be shifted by the number of Binders in between the declaration Binder and the location it’s referenced from.

Body

Fields

id: BodyId

s: &'a Scope<'a>

Lifetimes introduced by a fn are scoped to the call-site for that fn, if this is a fn body, otherwise the original definitions are used. Unspecified lifetimes are inferred, unless an elision scope is nested, e.g., (&T, fn(&T) -> &T); becomes (&'_ T, for<'a> fn(&'a T) -> &'a T).

ObjectLifetimeDefault

Fields

lifetime: Option<ResolvedArg>

s: &'a Scope<'a>

Use a specific lifetime (if Some) or leave it unset (to be inferred in a function body or potentially error outside one), for the default choice of lifetime in a trait object type.

Supertrait

Fields

bound_vars: Vec<BoundVariableKind>

s: &'a Scope<'a>

When we have nested trait refs, we concatenate late bound vars for inner trait refs from outer ones. But we also need to include any HRTB lifetimes encountered when identifying the trait that an associated type is declared on.

TraitRefBoundary

Fields

s: &'a Scope<'a>

AnonConstBoundary

Fields

s: &'a Scope<'a>

Disallows capturing non-lifetime binders from parent scopes.

This is necessary for something like for<T> [(); { /* references T */ }]:, since we don’t do something more correct like replacing any captured late-bound vars with early-bound params in the const’s own generics.

Root

Fields

opt_parent_item: Option<LocalDefId>