Struct rustc_hir_analysis::collect::ItemCtxt

pub struct ItemCtxt<'tcx> {
    tcx: TyCtxt<'tcx>,
    item_def_id: LocalDefId,
}

Context specific to some particular item. This is what implements AstConv.

ItemCtxt vs FnCtxt

ItemCtxt is primarily used to type-check item signatures and lower them from HIR to their ty::Ty representation, which is exposed using AstConv. It’s also used for the bodies of items like structs where the body (the fields) are just signatures.

This is in contrast to FnCtxt, which is used to type-check bodies of functions, closures, and consts – anywhere that expressions and statements show up.

An important thing to note is that ItemCtxt does no inference – it has no InferCtxt – while FnCtxt does do inference.

Trait predicates

ItemCtxt has information about the predicates that are defined on the trait. Unfortunately, this predicate information is available in various different forms at various points in the process. So we can’t just store a pointer to e.g., the AST or the parsed ty form, we have to be more flexible. To this end, the ItemCtxt is parameterized by a DefId that it uses to satisfy get_type_parameter_bounds requests, drawing the information from the AST (hir::Generics), recursively.

Fields
tcx: TyCtxt<'tcx>item_def_id: LocalDefId
Implementations

impl<'tcx> ItemCtxt<'tcx>

fn type_parameter_bounds_in_generics( &self, ast_generics: &'tcx Generics<'tcx>, param_def_id: LocalDefId, ty: Ty<'tcx>, filter: PredicateFilter ) -> Vec<(Clause<'tcx>, Span)>

Finds bounds from hir::Generics. This requires scanning through the AST. We do this to avoid having to convert all the bounds, which would create artificial cycles. Instead, we can only convert the bounds for a type parameter X if X::Foo is used.

fn bound_defines_assoc_item( &self, b: &GenericBound<'_>, assoc_name: Ident ) -> bool

impl<'tcx> ItemCtxt<'tcx>

pub fn new(tcx: TyCtxt<'tcx>, item_def_id: LocalDefId) -> ItemCtxt<'tcx>

pub fn to_ty(&self, ast_ty: &Ty<'_>) -> Ty<'tcx>

pub fn hir_id(&self) -> HirId

pub fn node(&self) -> Node<'tcx>