rustc_hir_analysis::hir_ty_lowering
Module cmse

Functions

is_valid_cmse_inputs 🔒

Returns whether the inputs will fit into the available registers

is_valid_cmse_output 🔒

Returns whether the output will fit into the available registers

should_emit_generic_error 🔒

validate_cmse_abi 🔒

Check conditions on inputs and outputs that the cmse ABIs impose: arguments and results MUST be returned via registers (i.e. MUST NOT spill to the stack). LLVM will also validate these conditions, but by checking them here rustc can emit nicer error messages.