Struct rustc_hir_analysis::structured_errors::WrongNumberOfGenericArgs

pub struct WrongNumberOfGenericArgs<'a, 'tcx> {
    pub(crate) tcx: TyCtxt<'tcx>,
    pub(crate) angle_brackets: AngleBrackets,
    pub(crate) gen_args_info: GenericArgsInfo,
    pub(crate) path_segment: &'a PathSegment<'a>,
    pub(crate) gen_params: &'a Generics,
    pub(crate) params_offset: usize,
    pub(crate) gen_args: &'a GenericArgs<'a>,
    pub(crate) def_id: DefId,
}

Handles the wrong number of type / lifetime / ... arguments family of error messages.

Fields
tcx: TyCtxt<'tcx>angle_brackets: AngleBracketsgen_args_info: GenericArgsInfopath_segment: &'a PathSegment<'a>

Offending path segment

gen_params: &'a Generics

Generic parameters as expected by type or trait

params_offset: usize

Index offset into parameters. Depends on whether Self is included and on number of lifetime parameters in case we’re processing missing or redundant type or constant arguments.

gen_args: &'a GenericArgs<'a>

Generic arguments as provided by user

def_id: DefId

DefId of the generic type

Implementations

impl<'a, 'tcx> WrongNumberOfGenericArgs<'a, 'tcx>

pub fn new(
tcx: TyCtxt<'tcx>,
gen_args_info: GenericArgsInfo,
path_segment: &'a PathSegment<'_>,
gen_params: &'a Generics,
params_offset: usize,
gen_args: &'a GenericArgs<'a>,
def_id: DefId
) -> Self

fn missing_lifetimes(&self) -> bool

fn kind(&self) -> &str

fn num_provided_args(&self) -> usize

fn num_provided_lifetime_args(&self) -> usize

fn num_provided_type_or_const_args(&self) -> usize

fn num_expected_lifetime_args(&self) -> usize

fn num_expected_type_or_const_args(&self) -> usize

fn num_expected_type_or_const_args_including_defaults(&self) -> usize

fn num_missing_lifetime_args(&self) -> usize

fn num_missing_type_or_const_args(&self) -> usize

fn num_excess_lifetime_args(&self) -> usize

fn num_excess_type_or_const_args(&self) -> usize

fn too_many_args_provided(&self) -> bool

fn not_enough_args_provided(&self) -> bool

fn get_lifetime_args_offset(&self) -> usize

fn get_num_default_params(&self) -> usize

fn is_synth_provided(&self) -> bool

fn get_quantifier_and_bound(&self) -> (&'static str, usize)

fn get_lifetime_args_suggestions_from_param_names(
&self,
path_hir_id: HirId,
num_params_to_take: usize
) -> String

fn get_type_or_const_args_suggestions_from_param_names(
&self,
num_params_to_take: usize
) -> String

fn get_unbound_associated_types(&self) -> Vec<String>

fn create_error_message(&self) -> String

fn start_diagnostics(&self) -> DiagnosticBuilder<'tcx, ErrorGuaranteed>

fn notify(&self, err: &mut Diagnostic)

Builds the expected 1 type argument / supplied 2 type arguments message.

fn suggest(&self, err: &mut Diagnostic)

fn suggest_adding_args(&self, err: &mut Diagnostic)

Suggests to add missing argument(s) when current invocation site already contains some generics:

type Map = HashMap<String>;

fn suggest_adding_lifetime_args(&self, err: &mut Diagnostic)

fn suggest_adding_type_and_const_args(&self, err: &mut Diagnostic)

fn suggest_moving_args_from_assoc_fn_to_trait(&self, err: &mut Diagnostic)

Suggests moving redundant argument(s) of an associate function to the trait it belongs to.

Into::into::<Option<_>>(42) // suggests considering `Into::<Option<_>>::into(42)`

fn suggest_moving_args_from_assoc_fn_to_trait_for_qualified_path(
&self,
err: &mut Diagnostic,
qpath: &'tcx QPath<'tcx>,
msg: String,
num_assoc_fn_excess_args: usize,
num_trait_generics_except_self: usize
)

fn suggest_moving_args_from_assoc_fn_to_trait_for_method_call(
&self,
err: &mut Diagnostic,
trait_def_id: DefId,
expr: &'tcx Expr<'tcx>,
msg: String,
num_assoc_fn_excess_args: usize,
num_trait_generics_except_self: usize
)

fn suggest_removing_args_or_generics(&self, err: &mut Diagnostic)

Suggests to remove redundant argument(s):

type Map = HashMap<String, String, String, String>;

fn show_definition(&self, err: &mut Diagnostic)

Builds the type defined here message.

fn note_synth_provided(&self, err: &mut Diagnostic)

Add note if impl Trait is explicitly specified.