rustc_hir_analysis::collect
Module resolve_bound_vars

Resolution of early vs late bound lifetimes.

Name resolution for lifetimes is performed on the AST and embedded into HIR. From this information, typechecking needs to transform the lifetime parameters into bound lifetimes. Lifetimes can be early-bound or late-bound. Construction of typechecking terms needs to visit the types in HIR to identify late-bound lifetimes and assign their Debruijn indices. This file is also responsible for assigning their semantics to implicit lifetimes in trait objects.

Structs

BoundVarContext ğŸ”’

NamedVarMap ğŸ”’

Maps the id of each bound variable reference to the variable decl that it corresponds to.

TruncatedScopeDebug ğŸ”’

Enums

BinderScopeType ğŸ”’

NonLifetimeBinderAllowed ğŸ”’

Scope ğŸ”’

Traits

RegionExt ğŸ”’

Functions

deny_non_region_late_bound ğŸ”’

is_late_bound_map ğŸ”’

Detects late-bound lifetimes and inserts them into late_bound.

late_arg_as_bound_arg ğŸ”’

object_lifetime_default ğŸ”’

provide ğŸ”’

resolve_bound_vars ğŸ”’

Computes the ResolveBoundVars map that contains data for an entire Item. You should not read the result of this query directly, but rather use named_variable_map, is_late_bound_map, etc.

Type Aliases

ScopeRef ğŸ”’