Module rustc_hir_analysis::collect::resolve_bound_vars

Resolution of early vs late bound lifetimes.

Name resolution for lifetimes is performed on the AST and embedded into HIR. From this information, typechecking needs to transform the lifetime parameters into bound lifetimes. Lifetimes can be early-bound or late-bound. Construction of typechecking terms needs to visit the types in HIR to identify late-bound lifetimes and assign their Debruijn indices. This file is also responsible for assigning their semantics to implicit lifetimes in trait objects.

Structs

BoundVarContext 🔒

NamedVarMap 🔒

Maps the id of each bound variable reference to the variable decl that it corresponds to.

TruncatedScopeDebug 🔒

Enums

BinderScopeType 🔒

NonLifetimeBinderAllowed 🔒

Scope 🔒

Traits

RegionExt 🔒

Functions

deny_non_region_late_bound

is_late_bound_map 🔒

Detects late-bound lifetimes and inserts them into late_bound.

late_arg_as_bound_arg 🔒

object_lifetime_default 🔒

provide 🔒

resolve_bound_vars 🔒

Computes the ResolveBoundVars map that contains data for an entire Item. You should not read the result of this query directly, but rather use named_variable_map, is_late_bound_map, etc.

Type Aliases

ScopeRef 🔒