rustc_hir_analysis::hir_ty_lowering
Module generics

Functions

check_generic_arg_count 🔒

Checks that the correct number of generic arguments have been provided. This is used both for datatypes and function calls.

check_generic_arg_count_for_call

Checks that the correct number of generic arguments have been provided. Used specifically for function calls.

generic_arg_mismatch_err 🔒

Report an error that a generic argument did not match the generic parameter that was expected.

lower_generic_args

Lower generic arguments from the HIR to the rustc_middle::ty representation.

prohibit_explicit_late_bound_lifetimes 🔒

Prohibits explicit lifetime arguments if late-bound lifetime parameters are present. This is used both for datatypes and function calls.