rustc_hir_analysis
Module variance

Module for inferring the variance of type and lifetime parameters. See the rustc dev guide chapter for more info.

Modules

constraints ğŸ”’

Code to gather up constraints. Constraint construction and representation

dump ğŸ”’

solve ğŸ”’

Code to solve constraints and write out the results. Constraint solving

terms ğŸ”’

Defines the TermsContext basically houses an arena where we can allocate terms.

xform ğŸ”’

Code for transforming variances.

Enums

ForceCaptureTraitArgs ğŸ”’

Functions

crate_variances ğŸ”’

provide ğŸ”’

variance_of_opaque ğŸ”’

variances_of ğŸ”’